requirements.txt
amqp==5.2.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
channels==4.0.0
channels-redis==4.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.1
exceptiongroup==1.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
kombu==5.3.4
msgpack==1.0.7
prompt-toolkit==3.0.43
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.25.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
